requirements.txt
altair==5.5.0
attrs==24.3.0
beautifulsoup4==4.12.3
blinker==1.9.0
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.1
cycler==0.12.1
dnspython==2.7.0
fake-useragent==1.5.1
fonttools==4.55.3
gitdb==4.0.12
GitPython==3.1.44
idna==3.10
Jinja2==3.1.5
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.8
langid==1.1.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
mdurl==0.1.2
narwhals==1.22.0
nltk==3.9.1
numpy==2.2.1
packaging==24.2
pandas==2.2.3
pillow==10.4.0
plotly==5.24.1
protobuf==5.29.3
pyarrow==18.1.0
pycountry==24.6.1
pydeck==0.9.1
Pygments==2.19.1
pymongo==4.10.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rpds-py==0.22.3
seaborn==0.13.2
six==1.17.0
smmap==5.0.2
soupsieve==2.6
streamlit==1.38.0
tenacity==8.5.0
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
wordcloud==1.9.3
